small_restructure: move unit tests
