+++ tests/resources/SSB_FDB/datastore/pending_operations.json
@@ -2,15 +2,9 @@
     "version": "0.0.0.0",
     "pendingOperations": [
         {
-            "datasetName" : "TEST_DATASET",
-            "operation" : "ADD",
-            "description" : "Nytt datasett om test",
-            "releaseStatus": "DRAFT"
-        },
-        {
-            "datasetName" : "ANOTHER_TEST_DATASET",
-            "operation" : "ADD",
-            "description" : "Enda et datasett om test",
+            "datasetName": "ANOTHER_TEST_DATASET",
+            "operation": "ADD",
+            "description": "Enda et datasett om test",
             "releaseStatus": "DRAFT"
         }
     ]
